brya: Fix battery wake-up logic
This fixes a circular dependency in waking up a battery after cut-off
(ship mode). Previously, we would report the battery as not present when
we failed to talk to the battery. This would prevent the charger task
from attempting to wake up the battery (i.e. enable pre-charge
current). Without pre-charge current, we can not talk to a cut-off
battery and would never be able to satisfy the conditions required to
turn on the battery.

The solution is to report BP_NOT_SURE back to the charger state machine
so that it enters the pre-charge state when we can not talk to the
battery.

BRANCH=none
BUG=b:180784200,b:181842699
TEST=cutoff battery from EC and reboot, observe battery wakes up.

Change-Id: Ib63167626dd7def685379fe718b3fadd043a341b
Signed-off-by: Caveh Jalali <caveh@chromium.org>