brya: Extend AP init-on duration
This increases the amount of time the power button task waits for the
system to stabilize before giving up on asserting the power button to
turn on the AP. When the battery is cut-off, it takes about 2 seconds
for the battery to be woken up and report a non-zero state-of-charge.
So, give the power button task 5 seconds to power on the AP instead of
the default 1 second.

BRANCH=none
BUG=b:180784200
TEST=cut-off battery and rebooted EC. noted AP is powered on after about
	2 seconds.

Change-Id: I6e590de8a9b38884376d90ef46e75d5467e729d6
Signed-off-by: Caveh Jalali <caveh@chromium.org>