TCPMv2: Do not apply CC Open when running without battery
This patch prevents applying CC Open when there is no battery.
On some boards EC RO doesn't save power-on reset flag in BBRAM when
EC reboot by Cr50 is expected. This behaviour was implemented in
CL:2255830. Brown-out reset flag is implemented only for LM4 chips.

Also we don't have access to information stored by TCPMv1 in BBRAM.
This functionality was removed from TCPMv2 in CL:2271002, so we can't
really determine if we should apply CC Open and from that time we are
applying CC Open always, except power-on and brown-out reset flags.
This patch adds missing battery to list of exceptions.

BUG=b:161775827
BRANCH=none
TEST=Flash EC ToT on Fleex. Disconnect battery physically and plug
     charger. Check if ChromeOS boots successfully.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Icc22fcce595693c744e1a55b905cdd58b5e7bfbd