stm32g4: ucpd: Move hbit clock config macros to .h file
Honeybuns needs to preset Rd on both CC lines while in RO. But, it
does not have a full usbc/pd stack. The ucpd driver file is not
included either. However, ucpd needs some basic initialization so that
Rd can be applied correctly. This CL moves the macros which are
required to configure the ucpd clocks to the .h file.

BUG=b:172493899
BRANCH=None
TEST=make BOARD=quiche

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I07614a941a05fa133ab4504d9241249066f41e29