driver: ln9310: use CC_I2C for logging
CC_CHIPSET is only available if the chipset task is enabled, but the
ln9310 has no other dependencies on power sequencing.  Bringup may
require this driver to be enabled before power sequencing.  It makes
more sense to log to the I2C channel, since we communicate with this
chip on I2C.

BUG=b:183054226
BRANCH=none
TEST=buildall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I3b040c88613be6075a834c6758b79965ab88fd76