CBI: Add CBI set to console command
Add the ability to set and remove tags to the CBI in the EC console.
This will allow boards which are not yet booting to the OS to set their
CBI fields from the EC directly.  It provides similar field size checks
to the ectool command version, though strings are more limited than with
ectool (less than 80 characters, versus over 200 possible with ectool).

BRANCH=None
BUG=b:183029344
TEST=on galith, use cbi to set and remove various fields

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ie7395e12467eeac4732cd17e923ee4a78a461b36