zephyr: lazor: Add initial USC-C config into the build
Add this file into the zephyr build along with the required CONFIG
options and device tree additions, so we can make a start on bringing
up USB-C and charging functionality.

Add all the required GPIOs here as well, so that things will work when
we move the code over in the following CLs.

BUG=b:183296099
BRANCH=none
TEST=build zephyr for lazor

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I8c4f54664d2d18b44ed8992e61e7b2f694f04f93