zephyr: Add support for hardware charge-ramping
Allow this option to be selected, turning it into a choice between the
two options. Make hardware ramping the default since it seems to be
preferred. It was apparently not supported on volteer only due to a
bug in the charger chip.

Also move the software-ramp requirement for ISL9241 to Kconfig. We
cannot 'select' a choice item like PLATFORM_EC_CHARGE_RAMP_SW in
PLATFORM_EC_CHARGER_ISL9241, so use a 'depends on' instead. Update
volteer to use software ramping since the default has changed.

BUG=b:183296099
BRANCH=none
TEST=build zephyr for lazor
build for volteer and check that CONFIG_PLATFORM_EC_CHARGE_RAMP_SW=y

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ic3ffd21cd10aa896a4c6dbe1d98d75776c89aefe