zephyr: volteer: Drop hard-coded battery enumeration
Now that this enum is available in a header file, include it in the
build, so it is available to any file that needs it. With ECOS the
enum is assumed to be present since it is declared in the board.h
header file.

This lets us tidy up the hard-coded enum in config_chip.h so we can
enable the battery on other boards.

Add the required config to volteer.

BUG=b:176121284
BRANCH=none
TEST=zmake configure -B /tmp/z/vol zephyr/projects/volteer -t zephyr &&
 zmake -L build  /tmp/z/vol

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I636410a38d156061efe67798c99a7d60c9fc6096