intel_x86: Assert RTCRST# in recovery mode
During microcode in-field firmware update, if the top swap is
enabled there is a chance that ucode is in staging area and the SPI
is corrupted. This will prevent the platform to boot properly.
To recover from this situation, recovery boot is triggered to reset
and disable the top swap so that the platform can boot from ucode in
RO thus RTC reset is added in EC code.

BUG=b:149547271
BRANCH=none
TEST=Tested on Volteer, RTCRTS# is asserted in recovery mode

Change-Id: Ie6354848dfb3cbc306d366969c03d0fb0a538955
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>