zephyr: lazor: enable PLATFORM_EC_PWM_DISPLIGHT
In Lazor, the display backlight is controlled by the EC chipset.
The kernel driver talks to EC over host command.

Enable PLATFORM_EC_PWM_DISPLIGHT such that the kernel driver
knows the PWM channel for the display backlight.

Can't test it yet as the host command interface is not ready.

BUG=b:182398910
BRANCH=None
TEST=zmake testall

Change-Id: I755ccabc7d27ff028a0b5f23df4b345f1bf8ca01
Signed-off-by: Wai-Hong Tam <waihong@google.com>