zephyr: i2c mapping using an enum
Currently, every project will need to provide a custom mapping header
for i2c as well as gpios. We'd like to move that overhead to the dts
files, which will make the board dts a lot more reusable as well as
clean up the bringup of new boards.

This change adds to the i2c_ports enum from the dts named i2c port
list. Each enum is only added if a node exists using that enum. This
also allowes for verifying that the enum is only used once.

BRANCH=none
BUG=b:177609422
TEST=build and flashed volteer
TEST=zmake testall

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I42cc098a9badac57a3781fa9dfaf32c0ec0c5aca