zephyr: i2c mapping using devicetree compatible strings
Currently, every project will need to provide a custom mapping header
for i2c as well as gpios. We'd like to move that overhead to the dts
files, which will make the board dts a lot more reusable as well as
clean up the bringup of new boards.

This change adds to the i2c_ports enum 2 entries per port. The first
is the compatible string which is used to identify the purpose of the
port prefixed with I2C_. The second is the name that is expected for
the port (this is the name that is references through the cros ec
common code).

When initializing the device we effectively reverse that mapping.
Unfortunately, there doesn't exist a devicetree _FOREACH like method
of iterating on these so both the common name and the compatible
string are needed (though this is only done once when adding a new
compatible string).

In this example (to keep review simple) I've only altered volteer's
I2C_PORT_POWER. For each new I2C_PORT_ value, we'll need to add a
new yaml file with the compatible string, declare the enum in the
i2c.h file, and initialize the device in the i2c.c file. Finally,
update the .dts files which should use it.

BRANCH=none
BUG=none
TEST=build and flashed volteer
TEST=zmake testall

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I42cc098a9badac57a3781fa9dfaf32c0ec0c5aca