TCPMv2: Treat errant PS_RDY as an indicator to turn Vconn off
When a partner sends PS_RDY after a Vconn swap and we believed our DUT
would be the new Vconn source, turn off our Vconn sourcing.  This should
hopefully allow us to get in sync with the partner's perspective on the
world.  If repeated Vconn swaps continue to result in the partner
sending PS_RDY as if they are the new source, then Vconn will remain off
on the connection.

Additionally, remove now unused error handling since this state is no
longer special-cased in pe_report_error.

BRANCH=None
BUG=b:182221344
TEST=on sasuke, confirm charger which sends PS_RDY incorrectly is able
to connect reliably

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ie7990b6018482f740969e0d0e0115a772e59355f