keyboard_scan: add a function to get refresh key row mask
In order to support the referesh key in row 2 and 3
in the same model, we need to create a function to get
the refresh key row mask at runtime. With this change,
we can override the function in the board layer and
return the correct referesh key row mask.

BUG=b:183353285
BRANCH=firmware-volteer-13672.B
TEST=TBD

Change-Id: Idf241958dbceca1c44234d78b177f34c8937a81d
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>