zmake: Use a common function to report a failure message
At present when zmake fails the resulting messages is independently
created in a different place for each subcommand.

Move this in a common function, so we can make it consistent.

Also adjust the output to show the command at the end, in case the
user wants to copy it out and try it. Drop the explanation mark since
errors are not all that exciting.

BUG=b:177096315
BRANCH=none
TEST=manually test by running zmake configure / build

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I913c3c86dcd257ed8a7f5d378b38e31d7c6d6465