zephyr: zmake: Support sequential builds
At present with 'zmake build' everything happens in parallel. This is
nice because it saves time.

However if there is a build error, or some warnings, these typically
appear twice, since both the RO and RW builds are (mostly?) compiling
the same code.

It is not useful to show the same message twice. As a partial
solution, add a -s flag to make the RO and RW builds happen one after
the other, stopping if the first one fails. This flag can be used in
an IDE to provide friendlier output.

Admittedly this results in a slower build, but the difference is only
material when Kconfig or device tree have changed and everything is
being rebuilt. Most of the time an incremental build only takes a
second, so this option is quite handy.

BUG=b:177096315
BRANCH=none
TEST=manually test by running zmake build / configure
FEATURE=test sudo -E emerge zephyr-build-tools

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ie66cf894a87b05a7648a77767e37e94636771c6a