raa489000: Add raa489000_enable_asgate()
This commit exports a function to allow the ASGATE to be enabled or
disabled when the charger IC is in the READY state.  This can help to
achieve a VBUS level of 0V if the READY state is required/desired.

BUG=b:183220414
BRANCH=dedede
TEST=With some other patches to use this function, build and flash on
madoo, plug in a charger, remove charger, verify that VBUS goes down
to less than 200mV and decays to 0.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Icb5316d9c206138e68fa4e55dee5b454182921e9