brya: Assert PROCHOT when VSYS droops too low
This addresses a problem where brya shuts down when running without a
battery. The root cause is that VSYS droops too much when SoC power
demand spikes.

The solution is to raise the VSYS_TH2 threshold in the bq25720 charger
chip to provide an early throttle feedback signal to the SoC during high
load. The default threshold is 5.9v which is low enough to cause
problems. On Brya, we see VSYS drop all the way down to 5.4v at which
point the SoC shuts down. The voltage should never be drawn down that
low, so tune the charger to assert PROCHOT at 7.0v instead. This
throttles the SoC sufficiently during high load to avoid
instability. 7.0v was selected after consulting with EE team.

BRANCH=none
BUG=b:183474163
TEST=booted brya to the OS without a battery

Change-Id: Id06e567b9f8a499da7ceb345e673e091c94d90c2
Signed-off-by: Caveh Jalali <caveh@chromium.org>