zephyr: driver: SHI: remove SHI version control
In NPCX7mnFA, the old SHI module and the new SHI module co-exist in this
chip. We need to set DEVALTF.bit7 to enable the new version of the SHI
module. In the latter npcx7 variants and npcx9 chips, only the new SHI
module is supported. Because NPCX7mnFA chip variants are not included in
the zephyr repo. We can remove the version control setttng.

BUG=b:182600858
BRANCH=none
TEST=Test host command "version" and "Hello" on npcx7_evb and the host
emulator.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I43b2fb5294f39e2bf04a0794c9d54e87c8e0ce33