zephyr: New test of ec app
Change zephy/app/ec/main.c:main to instead be named ec_app_main, so
that it can be called from tests, and created a trivial stub main to
call it. Fix up the CMake files for this change.

Add a new ec_app test which calls the ec_app_main() and verifies
that hooks of HOOK_INIT type are called. The CONFIG_PLATFORM_EC_HOOKS is
the only config currently enabled in tests, so that is the only thing
tested for now.

BUG=b:184273560
TEST=zmake configure --test zephyr/test/ec_app
BRANCH=none

Change-Id: I84ca9b36d3f387629ff5bd83463429d762cc2301
Signed-off-by: Jeremy Bettis <jbettis@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2799914
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>