zephyr: Update the temperature-sensor shim
At present this shim relies partially on a hard-coded mapping, in the
temp_sensors.c file. Also it uses the node IDs as enum values, which
makes it difficult to actually use the enums in the code.

With the ADC shim we need to be able to support using ADC enums in the
code, such as ADC_PSYS. This will force a change of naming in
enum adc_channel, which temperature sensors refer to.

Update the shim to use an enum for the temp-sensor ID and another for
the function to call to read the temperature.

BUG=b:175881324
BRANCH=none
TEST=build Zephyr for lazor and volteer

Change-Id: I08bd45568be525650b8527830053ee541d6240cf
Signed-off-by: Simon Glass <sjg@chromium.org>