lantis: Move C1 SM5803 processing to the PD_INT task
Since the SM5803 shares an interrupt line with the TCPC, allow the
PD_INT task for C1 to process interrupts for this chip.  This will
ensure that any interrupts from the charger are handled at a high
priority and cannot leave the shared IRQ line low for extended periods.

BUG=none
BRANCH=dedede
TEST=On lantis, confirm charger attach to C1 works reliably.

Signed-off-by: Tommy Chung <tommy.chung@quanta.corp-partner.google.com>
Change-Id: Id363b743eb7825f4a03f7119248218d0259e9d5b