zephyr: fix initialization priority to be forward compatible
In upstream Zephyr the initialization of the GPIO module (and most other
drivers) was changed. This breaks our gpio and in return also flash
initialization.

Update both gpio and flash initialization to work with Kconfig and add
a compile time validation to be able to catch this sooner in the future.

BRANCH=none
BUG=none
TEST=zmake testall

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I44e5867dbf4cb3d5934bf0d0807dcc1aa6c778e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2808335
Reviewed-by: Simon Glass <sjg@chromium.org>