zephyr: zmake: Add a test for output filtering
Test that the filtering does what we expect. This works by using some
pre-canned output for the RO and RW, running each file through a
separate 'cat' command in its own process, then comparing the results
obtained in the log.

In creating this test, unexpected behaviour was found with the
image-size logging (Memory, FLASH:, etc), in that they end up on the
ERROR level, if previous output has appeared on that level. This is
because the default log level changes to whatever the last log output
was. So fix this at the same time.

For now we don't check the DEBUG level, since it has a bug.

BUG=b:177096315
BRANCH=none
TEST=(cd zephyr/zmake/ && python3 -m pytest  .)

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Iaa777815b577ab814eb2e69d34a3047ae57f0b45