zephyr: zmake: Avoid using pip3 for install
When running on a distribution it seems better to use the standard
install procedure if possible, with pip3 reserved for things that are
not available. If the user has an old distribution that will presumably
fall back to using pip3 anyway.

Update to use 'apt-get install' in the docs.

BUG=b:181253613
BRANCH=none
TEST=check it in gitiles

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I539518a3009e73cb948145ca68f868fa5588909d