brya: Update EC GPIOs for new board rev.
This updates the GPIO functional changes to support the new rev. of the
brya board.

EC_KB_BL_EN moved from GPIO85 to GPIOA3 (was EN_SLP_Z). EC_KB_BL_EN is
set to low since that's the preferred setting for the initial keyboard
backlight. This does not break the old board due to a hardware bug on
the old board that causes it to come out of sleep unconditionally.
GPIO85 does not need to be configured here when PSL is enabled.
EN_SLP_Z is no longer used.

EC_RST_ODL (VCC1_RST#/GPO77) has been removed from config. This was an
unused input on the old board. This pin functions as a wake source when
we enable HIBERNATE_PSL_VCC1_RST_WAKEUP and does not need to be
configured here.

finally, PSL is enabled on pins PSL_IN1-3.

BRANCH=none
BUG=b:184811017,b:183246197
TEST=both old new new revison of the brya board can boot

Cq-Depend: chromium:2728679
Change-Id: I4b3ab17bb44d18167328faee1a1b604bf0428dd3
Signed-off-by: Caveh Jalali <caveh@chromium.org>