i2c_controller: zephyr: explicitly strip flags from addr_flags
The ECOS I2C APIs use addr_flag to pass the I2C address together with
few flags. These are currently passed to the Zephyr API stripping only
the PEC bit, and no other flags are used in Zephyr platforms at the
moment, but if any flag would be present it would confuse the Zephyr I2C
stack.

This changes the Zephyr part of the driver to explicitly strip any
possible flag from add_flag, and also add a warning to signal that we
are discarding some information.

BUG=none
BRANCH=none
TEST=build and flash on volteer
TEST=cbi shell command

Change-Id: I497fe33c52c78d783c86711928067a4aa0a321d7
Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>