zephyr: lazor: Add LED support
Add GPIOs to control the LED and add board/lazor/led.c to the compiled
list.

The color of the LED will be changing when the charging is supported.

BUG=b:182398910
BRANCH=none
TEST=Verfiy the LED color according to the charging state. If the
charging is not supported, disable led_set_color function and set the
color manually with gpioset GPIO_EC_CHG_LED_Y/B_C1 0/1.

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I910c58efa6b3771bacb264398acdcf171071f04a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2817478
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>