Power: Add configurable S5 timeout
Similar to the hibdelay command, allow the time we spend in S5 to be
configurable for the sake of power testing in this state.  It may be set
to 0 for an immediate drop, or -1 to stay in S5 forever.

BRANCH=None
BUG=b:184941436
TEST=on guybrush, confirm setting 0 causes an immediate transition,
setting 50 waits 50 seconds, and -1 appears to stay in S5 until the
power button is pressed

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: If2f8cf91ba982822e2e33cb17f84f38c40907d2e