zephyr: zmake: bind lambda for pytest using default argument
In Python, a lambda or function will capture variables in the parent
scope, but using a variable in a loop does not create a new scope.

Thus, the test_file argument is being executed as the current value of
test_file, not necessarily the value that happened that iteration in
the loop.

Prior to this CL, we were not guaranteed to run every pytest.  It's
racy and causes us to run some tests twice or even three times, while
completely skipping other tests.

Using a default argument in the lambda will bind to the current value
instead of the variable.

See this SO post for further discussion:
https://stackoverflow.com/questions/2295290

BUG=none
BRANCH=none
TEST=zmake testall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I5eda075980cce0af13fb7a114fd1917fa6732754
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2823148
Reviewed-by: Yuval Peress <peress@chromium.org>