TCPMv2: PE: Refer to vconn swap policy for sop'/sop''
There is a vconn swap check in pe_attempt_discovery that is used to
trigger a possible vconn role swap requeest. This check takes into
account a board's specific policy (if necessary). But, a vconn role
swap can also be triggered when sending a VDM message to
SOP'/SOP''. In this case, board specific policy still needs to be
checked.

BUG=b:183026242
BRANCH=quiche
TEST=tested on quiche and verified that there is no longer a continous
stream of vconn swap attempts.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I5d239c78dc0d324e0c3dbb99dfc7fa093e2f6570