honeybuns: Modify vconn and dr swap policy
This CL modifies honeybuns specific policy for data role and vconn
swaps. For vconn swaps, allow if PE request flag is set and we are
currently not the vconn source. For data role swaps, add a check to
only allow if the flag for check data role in PE is set. Otherwise, if
the port partner rejects the data role swap, we can end up sending a
non-stop stream of data role swap request messages.

BUG=b:183026242
BRANCH=quiche
TEST=tested on quiche
For vconn swap, verified that we don't end up doing multiple vconn
swaps back and forth with the host port.
For data role swap, verfied that even when the data role swap is
rejected, there is still just one attempt.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ic3e96d7aeac40c9f8c9bf97c5f4b0142ff8365f4