flash_ec: introduce early Ti50 support
For the purposes of UUT based EC programming Ti50 is very close to
Cr50, the only difference is that Ti50 does not allow enabling UUT
mode before EC reset is asserted.

This patch makes sure that processing path for Ti50 and Cr50 is the
same, and adds asserting of the EC reset before UUT programming mode
is enabled.

Also done some clean up, declaring variables 'local' in functions this
patch touches.

BRANCH=none
BUG=b:184770575
TEST=the procedure is as follows, inside chrome OS chroot:

 $ sudo servod --board brya -c ccd_ti50.xml

   In a different terminal in the EC tree:

 $ make BOARD=brya
 $ ./util/flash_ec --board=brya

  Verify that the EC is up and running the latest image.

  Also verified Cr50 CCD operation by programming a bobba EC image:

 $ sudo servod --board bobba

   In a different terminal in the EC tree:

 $ # place a booba image into build/bobba/ec.bin
 $ ./util/flash_ec --board=bobba

  Verify that the bobba device was programmed using UUT and is running
  the updated EC image.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ib266078c66586591d0115fe76214ae8c8921a547