WIP: ps8815: Add FRS support
This enables FRS (fast role swap) support on the ps8815. Also, support
is added for an optional feature to control FRS using a GPIO on the
ps8815. This would typically control FRS on the port's PPC. Since this
feature may not be available or needed on all ps8815 TCPCs, a TCPC port
flag is used to enable this feature on desired ports.

BRANCH=none
BUG=b:183127346

Change-Id: I949958e51512b096e55d405de9fc7db848e58ceb
Signed-off-by: Caveh Jalali <caveh@chromium.org>